Crate harfrust

A complete harfbuzz shaping algorithm port to Rust.

Modules

script

Predefined scripts.

Structs

BufferFlags

Flags for buffers.

Feature

A feature tag with an accompanying range specifying on which subslice of shapes input it should be applied.

FontRef

Reference to an in-memory font.

GlyphBuffer

A buffer that contains the results of the shaping process.

GlyphInfo

A glyph info.

GlyphPosition

Holds the positions of the glyph in both horizontal and vertical directions.

Language

A language tag.

Script

A text script.

SerializeFlags

Flags used for serialization with a BufferSerializer.

ShapePlan

A reusable plan for shaping a text buffer.

ShapePlanKey

A key used for selecting a shape plan.

Shaper

A configured shaper.

ShaperBuilder

Builder type for constructing a Shaper.

ShaperData

Data required for shaping with a single font.

ShaperInstance

An instance of a variable font.

Tag

An OpenType tag.

UnicodeBuffer

A buffer that contains an input string ready for shaping.

Variation

A font variation.

Enums

BufferClusterLevel

A cluster level.

Direction

Defines the direction in which text is to be read.

Type Aliases

NormalizedCoord

Type alias for a normalized variation coordinate.