harfrust::font

Struct FontInstance

pub(crate) struct FontInstance {
    font: Font,
    size: Option<f32>,
    coords: CoordStorage,
    feature_vars: FeatureVarsStorage,
}

A specific instance of a font, with a size and variation settings.

Fields
font: Fontsize: Option<f32>coords: CoordStoragefeature_vars: FeatureVarsStorage
Implementations

impl FontInstance

pub fn builder(font: &Font) -> FontInstanceBuilder

Returns a builder for configuring a font instance from the given font.

pub fn font(&self) -> &Font

Returns the font for this instance.

pub fn size(&self) -> Option<f32>

Returns the size of the font instance, in pixels per em.

pub fn normalized_coords(&self) -> &[F2Dot14]

Returns the normalized variation coordinates for this font instance.

pub fn feature_variations(&self) -> FontFeatureVariations

Returns the selected feature variations for this font instance.