harfrust::hb

Module buffer

Macros

declare_buffer_var 🔒

declare_buffer_var_alias 🔒

foreach_cluster 🔒

foreach_grapheme 🔒

foreach_group 🔒

foreach_syllable 🔒

Structs

GlyphBuffer

A buffer that contains the results of the shaping process.

GlyphFlags

Flags that describe the properties of a glyph.

GlyphInfo

A glyph info.

GlyphPosition

Holds the positions of the glyph in both horizontal and vertical directions.

GlyphPropsFlags

UnicodeBuffer

A buffer that contains an input string ready for shaping.

UnicodeProps

buffer_var_shape 🔒

hb_buffer_t

Constants

CONTEXT_LENGTH 🔒

HB_BUFFER_CLUSTER_LEVEL_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_DEFAULT

HB_BUFFER_CLUSTER_LEVEL_GRAPHEMES

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_CHARACTERS

HB_BUFFER_CLUSTER_LEVEL_MONOTONE_GRAPHEMES

HB_BUFFER_SCRATCH_FLAG_DEFAULT

HB_BUFFER_SCRATCH_FLAG_HAS_BROKEN_SYLLABLE

HB_BUFFER_SCRATCH_FLAG_HAS_CGJ

HB_BUFFER_SCRATCH_FLAG_HAS_CONTINUATIONS

HB_BUFFER_SCRATCH_FLAG_HAS_DEFAULT_IGNORABLES

HB_BUFFER_SCRATCH_FLAG_HAS_FRACTION_SLASH

HB_BUFFER_SCRATCH_FLAG_HAS_GPOS_ATTACHMENT

HB_BUFFER_SCRATCH_FLAG_HAS_SPACE_FALLBACK

HB_BUFFER_SCRATCH_FLAG_HAS_VARIATION_SELECTOR_FALLBACK

HB_BUFFER_SCRATCH_FLAG_SHAPER0

Traits

SerializerFont

Functions

_cluster_group_func 🔒

Type Aliases

hb_buffer_cluster_level_t

hb_buffer_scratch_flags_t