harfrust::hb::buffer

Struct GlyphInfo

#[repr(C)]
pub struct GlyphInfo {
    pub glyph_id: u32,
    pub(crate) mask: u32,
    pub cluster: u32,
    pub(crate) vars: [u32; 2],
}

A glyph info.

Structure that holds information about the glyphs and their relation to input text.

HarfBuzz calls this hb_glyph_info_t. See the documentation and source.

Fields
glyph_id: u32

A selected glyph.

Guarantee to be <= u16::MAX.

mask: u32cluster: u32

An index to the start of the grapheme cluster in the original string.

Read more on clusters.

vars: [u32; 2]
Implementations

impl GlyphInfo

pub const fn flags(&self) -> GlyphFlags

Returns the flags for this glyph.

pub fn unsafe_to_break(&self) -> bool

Indicates that if input text is broken at the beginning of the cluster this glyph is part of, then both sides need to be re-shaped, as the result might be different.

On the flip side, it means that when this flag is not present, then it’s safe to break the glyph-run at the beginning of this cluster, and the two sides represent the exact same result one would get if breaking input text at the beginning of this cluster and shaping the two sides separately. This can be used to optimize paragraph layout, by avoiding re-shaping of each line after line-breaking, or limiting the reshaping to a small piece around the breaking point only.

pub fn unsafe_to_concat(&self) -> bool

Indicates that if input text is changed on one side of the beginning of the cluster this glyph is part of, then the shaping results for the other side might change. Note that the absence of this flag will NOT by itself mean that it IS safe to concat text. Only two pieces of text both of which clear of this flag can be concatenated safely.

This can be used to optimize paragraph layout, by avoiding re-shaping of each line after line-breaking, by limiting the reshaping to a small piece around the breaking position only, even if the breaking position carries the unsafe-to-break flag or when hyphenation or other text transformation happens at line-break position, in the following way:

Iterate back from the line-break position until the first cluster start position that is NOT unsafe-to-concat.

Shape the segment from there till the end of line.

Check whether the resulting glyph-run also is clear of the unsafe-to-concat at its start-of-text position; if it is, just splice it into place and the line is shaped; If not, move on to a position further back that is clear of unsafe-to-concat and retry from there, and repeat.

At the start of next line a similar algorithm can be implemented. That is:

Iterate forward from the line-break position until the first cluster start position that is NOT unsafe-to-concat.

Shape the segment from beginning of the line to that position.

Check whether the resulting glyph-run also is clear of the unsafe-to-concat at its end-of-text position; if it is, just splice it into place and the beginning is shaped; If not, move on to a position further forward that is clear of unsafe-to-concat and retry up to there, and repeat. A slight complication will arise in the implementation of the algorithm above, because while our buffer API has a way to return flags for position corresponding to start-of-text, there is currently no position corresponding to end-of-text. This limitation can be alleviated by shaping more text than needed and looking for unsafe-to-concat flag within text clusters. The unsafe-to-break flag will always imply this flag. To use this flag, you must enable the buffer flag BufferFlags::PRODUCE_UNSAFE_TO_CONCAT during shaping, otherwise the buffer flag will not be reliably produced.

pub fn safe_to_insert_tatweel(&self) -> bool

In scripts that use elongation (Arabic, Mongolian, Syriac, etc.), this flag signifies that it is safe to insert a U+0640 TATWEEL character before this cluster for elongation. This flag does not determine the script-specific elongation places, but only when it is safe to do the elongation without interrupting text shaping.

pub(crate) fn as_codepoint(&self) -> u32

pub(crate) fn as_glyph(&self) -> GlyphId

pub(crate) fn as_gid16(&self) -> Option<GlyphId16>

pub(crate) fn init_unicode_props(&mut self, scratch_flags: &mut u32)

pub(crate) fn unhide(&mut self)

impl GlyphInfo

pub(crate) const GLYPH_PROPS_VAR: buffer_var_shape

pub(crate) const LIG_PROPS_VAR: buffer_var_shape

pub(crate) const SYLLABLE_VAR: buffer_var_shape

pub(crate) const UNICODE_PROPS_VAR: buffer_var_shape

pub(crate) fn glyph_props(&self) -> u16

pub(crate) fn set_glyph_props(&mut self, value: u16)

pub(crate) fn lig_props(&self) -> u8

pub(crate) fn set_lig_props(&mut self, value: u8)

pub(crate) fn syllable(&self) -> u8

pub(crate) fn set_syllable(&mut self, value: u8)

pub(crate) fn unicode_props(&self) -> u16

pub(crate) fn set_unicode_props(&mut self, value: u16)

impl GlyphInfo

pub(crate) fn set_general_category(&mut self, gen_cat: GeneralCategory)

HB: _hb_glyph_info_set_general_category

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L260

pub(crate) fn general_category(&self) -> GeneralCategory

HB: _hb_glyph_info_get_general_category

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L268

pub(crate) fn is_unicode_mark(&self) -> bool

HB: _hb_glyph_info_is_unicode_mark

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L274

pub(crate) fn set_modified_combining_class(&mut self, modified_class: u8)

HB: _hb_glyph_info_set_modified_combining_class

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L279

pub(crate) fn modified_combining_class(&self) -> u8

HB: _hb_glyph_info_get_modified_combining_class

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L287

pub(crate) fn is_unicode_space(&self) -> bool

HB: _hb_glyph_info_is_unicode_space

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L294

pub(crate) fn set_unicode_space_fallback_type(&mut self, s: u8)

HB: _hb_glyph_info_set_unicode_space_fallback_type

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L300

pub(crate) fn unicode_space_fallback_type(&self) -> u8

HB: _hb_glyph_info_get_unicode_space_fallback_type

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L307

pub(crate) fn is_variation_selector(&self) -> bool

HB: _hb_glyph_info_is_variation_selector

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L314

pub(crate) fn set_variation_selector(&mut self, customize: bool)

HB: _hb_glyph_info_set_variation_selector

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L321

pub(crate) fn is_default_ignorable(&self) -> bool

HB: _hb_glyph_info_is_default_ignorable

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L338

pub(crate) fn _set_default_ignorable(&mut self)

HB: _hb_glyph_info_set_default_ignorable

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L344

pub(crate) fn clear_default_ignorable(&mut self)

HB: _hb_glyph_info_clear_default_ignorable

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L349

pub(crate) fn is_hidden(&self) -> bool

HB: _hb_glyph_info_is_hidden

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L354

pub(crate) fn set_continuation(&mut self, scratch_flags: &mut u32)

HB: _hb_glyph_info_set_continuation

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L365

pub(crate) fn clear_continuation(&mut self)

HB: _hb_glyph_info_clear_continuation

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L370

pub(crate) fn is_continuation(&self) -> bool

HB: _hb_glyph_info_is_continuation

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L375

impl GlyphInfo

const IS_LIG_BASE: u8 = 0x10

pub(crate) fn is_unicode_format(&self) -> bool

HB: _hb_glyph_info_is_unicode_format

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L398

pub(crate) fn is_zwnj(&self) -> bool

HB: _hb_glyph_info_is_zwnj

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L404

pub(crate) fn is_zwj(&self) -> bool

HB: _hb_glyph_info_is_zwj

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L409

pub(crate) fn is_aat_deleted(&self) -> bool

HB: _hb_glyph_info_is_aat_deleted

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L426

pub(crate) fn set_aat_deleted(&mut self)

HB: _hb_glyph_info_set_aat_deleted

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L431

pub(crate) fn set_lig_props_for_ligature( &mut self, lig_id: u8, lig_num_comps: u8, )

HB: _hb_glyph_info_set_lig_props_for_ligature

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L472

pub(crate) fn set_lig_props_for_mark(&mut self, lig_id: u8, lig_comp: u8)

HB: _hb_glyph_info_set_lig_props_for_mark

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L480

pub(crate) fn set_lig_props_for_component(&mut self, comp: u8)

HB: _hb_glyph_info_set_lig_props_for_component

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L488

pub(crate) fn lig_id(&self) -> u8

HB: _hb_glyph_info_get_lig_id

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L494

pub(crate) fn ligated_internal(&self) -> bool

HB: _hb_glyph_info_ligated_internal

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L500

pub(crate) fn lig_comp(&self) -> u8

HB: _hb_glyph_info_get_lig_comp

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L506

pub(crate) fn lig_num_comps(&self) -> u8

HB: _hb_glyph_info_get_lig_num_comps

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L515

pub(crate) fn is_base_glyph(&self) -> bool

HB: _hb_glyph_info_is_base_glyph

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L548

pub(crate) fn is_ligature(&self) -> bool

HB: _hb_glyph_info_is_ligature

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L554

pub(crate) fn is_mark(&self) -> bool

HB: _hb_glyph_info_is_mark

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L560

pub(crate) fn substituted(&self) -> bool

HB: _hb_glyph_info_substituted

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L566

pub(crate) fn ligated(&self) -> bool

HB: _hb_glyph_info_ligated

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L572

pub(crate) fn multiplied(&self) -> bool

HB: _hb_glyph_info_multiplied

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L578

pub(crate) fn ligated_and_didnt_multiply(&self) -> bool

HB: _hb_glyph_info_ligated_and_didnt_multiply

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L584

pub(crate) fn clear_ligated_and_multiplied(&mut self)

HB: _hb_glyph_info_clear_ligated_and_multiplied

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L590

pub(crate) fn clear_substituted(&mut self)

HB: _hb_glyph_info_clear_substituted

See https://github.com/harfbuzz/harfbuzz/blob/368598b5bd9c37a15cb0fd5438b8e617e254609b/src/hb-ot-layout.hh#L597

impl GlyphInfo

pub(crate) const NORMALIZER_GLYPH_INDEX_VAR: buffer_var_shape

pub(crate) fn normalizer_glyph_index(&self) -> u32

pub(crate) fn set_normalizer_glyph_index(&mut self, value: u32)

impl GlyphInfo

pub(crate) const OT_SHAPER_VAR_U8_CATEGORY_VAR: buffer_var_shape

pub(crate) const OT_SHAPER_VAR_U8_AUXILIARY_VAR: buffer_var_shape

pub(crate) fn ot_shaper_var_u8_category(&self) -> u8

pub(crate) fn set_ot_shaper_var_u8_category(&mut self, value: u8)

pub(crate) fn ot_shaper_var_u8_auxiliary(&self) -> u8

pub(crate) fn set_ot_shaper_var_u8_auxiliary(&mut self, value: u8)

impl GlyphInfo

pub(crate) const ARABIC_SHAPING_ACTION_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_AUXILIARY_VAR

pub(crate) fn arabic_shaping_action(&self) -> u8

pub(crate) fn set_arabic_shaping_action(&mut self, value: u8)

impl GlyphInfo

pub(crate) const HANGUL_SHAPING_FEATURE_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_AUXILIARY_VAR

pub(crate) fn hangul_shaping_feature(&self) -> u8

pub(crate) fn set_hangul_shaping_feature(&mut self, value: u8)

impl GlyphInfo

pub(crate) const INDIC_CATEGORY_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_CATEGORY_VAR

pub(crate) const INDIC_POSITION_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_AUXILIARY_VAR

pub(crate) fn indic_category(&self) -> u8

pub(crate) fn set_indic_category(&mut self, value: u8)

pub(crate) fn indic_position(&self) -> u8

pub(crate) fn set_indic_position(&mut self, value: u8)

fn is_one_of(&self, flags: u32) -> bool

fn is_joiner(&self) -> bool

pub(crate) fn is_consonant(&self) -> bool

fn is_halant(&self) -> bool

fn set_indic_properties(&mut self)

impl GlyphInfo

pub(crate) const KHMER_CATEGORY_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_CATEGORY_VAR

pub(crate) fn khmer_category(&self) -> u8

pub(crate) fn set_khmer_category(&mut self, value: u8)

impl GlyphInfo

fn set_khmer_properties(&mut self)

impl GlyphInfo

pub(crate) const MYANMAR_CATEGORY_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_CATEGORY_VAR

pub(crate) const MYANMAR_POSITION_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_AUXILIARY_VAR

pub(crate) fn myanmar_category(&self) -> u8

pub(crate) fn set_myanmar_category(&mut self, value: u8)

pub(crate) fn myanmar_position(&self) -> u8

pub(crate) fn set_myanmar_position(&mut self, value: u8)

impl GlyphInfo

fn set_myanmar_properties(&mut self)

impl GlyphInfo

pub(crate) const USE_CATEGORY_VAR: buffer_var_shape = GlyphInfo::OT_SHAPER_VAR_U8_CATEGORY_VAR

pub(crate) fn use_category(&self) -> u8

pub(crate) fn set_use_category(&mut self, value: u8)

fn is_halant_use(&self) -> bool