harfrust::hb

Module common

Modules

script

Predefined scripts.

Structs

Feature

A feature tag with an accompanying range specifying on which subslice of shapes input it should be applied.

Language

A language tag.

Script

A text script.

Variation

A font variation.

Enums

Direction

Defines the direction in which text is to be read.

Constants

HB_FEATURE_GLOBAL_END

HB_FEATURE_GLOBAL_START

Traits

TagExt

Type Aliases

SmallVecLanguage 🔒