harfrust::hb::font_funcs

Struct BuiltinFontFuncs

pub struct BuiltinFontFuncs<'a> {
    face: &'a hb_font_t<'a>,
    glyph_metrics: OnceCell<GlyphMetrics<'a>>,
    charmap: OnceCell<Charmap<'a>>,
}

Default implementations backed by font tables.

Fields
face: &'a hb_font_t<'a>glyph_metrics: OnceCell<GlyphMetrics<'a>>charmap: OnceCell<Charmap<'a>>
Implementations

impl<'a> BuiltinFontFuncs<'a>

pub(crate) fn new(face: &'a hb_font_t<'a>) -> Self

fn coords(&self) -> &[F2Dot14]

fn charmap(&self) -> &Charmap<'a>

fn glyph_metrics(&self) -> &GlyphMetrics<'a>

pub fn nominal_glyph(&self, c: u32) -> Option<GlyphId>

Maps a Unicode scalar value to a nominal glyph.

pub fn variant_glyph(&self, c: u32, vs: u32) -> Option<GlyphId>

Maps a Unicode scalar value and variation selector to a glyph.

pub fn advance_width(&self, glyph: GlyphId) -> i32

Returns the horizontal advance for a glyph.

pub fn advance_height(&self, glyph: GlyphId) -> i32

Returns the vertical advance for a glyph.

pub fn vertical_origin(&self, glyph: GlyphId) -> (i32, i32)

Returns the vertical origin for a glyph.

pub fn extents(&self, glyph: GlyphId) -> Option<GlyphExtents>

Returns extents for a glyph if available.

pub fn populate_advance_widths(&self, batch: AdvanceWidthBatch<'_>)

Populates horizontal advances for all entries in the batch.